+++ CHANGES.rst
@@ -6,9 +6,14 @@ Version: 1.2.3a19 (unreleased)
 
 IMPROVEMENT:
 
+  * Dry-run mode: Detects now undefined steps.
   * SummaryReporter: Summary shows untested items if one ore more exist.
   * issue #103: sort feature file by name in a given directory (provided by: gurneyalex).
   * issue  #42: Show all undefined steps taking tags into account (provided by: roignac, jenisys)
+
+CHANGES:
+
+  * Dry-run mode: Uses untested counts now (was using: skipped counts).
 
 FIXED: